Autosynchronization between users in terms of lists and cards

See merge request cse1105/2022-2023/teams/oopp-team-62!61

+++ server/src/main/java/server/api/CardSavingController.java
@@ -48,6 +48,11 @@ public class CardSavingController {
 
     @DeleteMapping(path = {"delete/{id}"})
     public void delete(@PathVariable long id) {
+        Card card = repo.findById(id).orElse(null);
+        if (card == null) {
+            return;
+        }
+        msgs.convertAndSend("/topic/card", card);
         repo.deleteById(id);
     }
